refactor(plugin-exercise): move interactive ex selection, cleanup tsx
